Test : 테스트코드 오류 수정 및 BeforeEach 추가
